+++ fabric/src/main/java/com/lx862/jcm/mod/config/ClientConfig.java
@@ -10,7 +10,6 @@ public class ClientConfig extends Config {
 
     private static final Path CONFIG_PATH = MinecraftClient.getInstance().getRunDirectoryMapped().toPath().resolve("config").resolve("jsblock_client.json");
     public boolean disableRendering;
-    public boolean useCustomFont;
     public boolean debug;
     public boolean useNewTextRenderer;
 
@@ -21,7 +20,6 @@ public class ClientConfig extends Config {
     public void fromJson(JsonObject jsonConfig) {
         JCMLogger.info("Loading client config...");
         this.disableRendering = jsonConfig.get("disable_rendering").getAsBoolean();
-        this.useCustomFont = jsonConfig.get("custom_font").getAsBoolean();
         this.debug = jsonConfig.get("debug_mode").getAsBoolean();
         this.useNewTextRenderer = jsonConfig.get("new_text_renderer").getAsBoolean();
     }
@@ -30,7 +28,6 @@ public class ClientConfig extends Config {
         JCMLogger.info("Writing client config...");
         final JsonObject jsonConfig = new JsonObject();
         jsonConfig.addProperty("disable_rendering", disableRendering);
-        jsonConfig.addProperty("custom_font", useCustomFont);
         jsonConfig.addProperty("debug_mode", debug);
         jsonConfig.addProperty("new_text_renderer", useNewTextRenderer);
         return jsonConfig;